+++ bc/recruitment/models.py
@@ -6,10 +6,13 @@ from django.db.models import Count, F
 from django.db.models.signals import pre_save
 from django.dispatch import receiver
 from django.shortcuts import get_object_or_404, render
+from django.template.loader import render_to_string
+from django.urls import reverse
 from django.utils.functional import cached_property
 from django.utils.text import slugify
 
 from wagtail.admin.edit_handlers import FieldPanel, MultiFieldPanel, StreamFieldPanel
+from wagtail.admin.mail import send_mail
 from wagtail.contrib.routable_page.models import RoutablePageMixin, route
 from wagtail.core import blocks
 from wagtail.core.fields import StreamField
@@ -60,6 +63,7 @@ class JobCategory(models.Model):
 
     get_subcategories_list.short_description = "Subcategories"
 
+    @staticmethod
     def get_categories_summary():
         """Returns a QuerySet that returns dictionaries, when used as an iterable.
 
@@ -279,18 +283,40 @@ class RecruitmentIndexPage(BasePage):
 
 class JobAlertSubscription(models.Model):
     email = models.EmailField()
-    search = models.CharField(
-        max_length=255
-    )  # Serialised list of selected category ids and search term
+    search = models.TextField(
+        default="{}", editable=False
+    )  # stop site admins from entering bad values
     confirmed = models.BooleanField(default=False)
     created = models.DateTimeField(auto_now_add=True)
     token = models.CharField(max_length=255, unique=True, editable=False)
+
+    @property
+    def confirmation_url(self):
+        return reverse("confirm_job_alert", args=[self.token])
+
+    @property
+    def unsubscribe_url(self):
+        return reverse("unsubscribe_job_alert", args=[self.token])
 
     def full_clean(self, *args, **kwargs):
         if not self.token:
             self.token = secrets.token_urlsafe(32)
 
         super().full_clean(*args, **kwargs)
+
+    def send_confirmation_email(self, request):
+        template_name = "patterns/email/confirm_job_alert.html"
+        context = {}
+        context[
+            "search"
+        ] = self.search  # TODO: add search summary in human readable format
+        context["confirmation_url"] = request.build_absolute_uri(self.confirmation_url)
+        context["unsubscribe_url"] = request.build_absolute_uri(self.unsubscribe_url)
+
+        content = render_to_string(template_name, context=context)
+        send_mail(
+            "Job alert subscription", content, [self.email],
+        )
 
 
 class JobAlertNotificationTask(models.Model):