+++ bc/recruitment/tests/test_search.py
@@ -1,0 +1,13 @@
+# from bc.recruitment.utils import is_recruitment_site, get_current_search, get_job_search_results
+# from bc.recruitment.constants import JOB_FILTERS
+# from bc.recruitment.models import JobAlertSubscription
+
+# TODO: Add tests
+#
+#   test JobAlertSubscription correctly sets token
+#   test JobAlertSubscription.send_confirmation_email()
+#   test is_recruitment_site()
+#   test get_current_search() works for different filters and empty filters
+#   test get_job_search_results() works for different filters and empty filters
+
+# TODO: Test job alert subscription view