+++ bc/recruitment_api/tests/test_search.py
@@ -1,4 +1,0 @@
-# from bc.recruitment.utils import is_recruitment_site, get_jobs_search_results
-
-# test utils methods
-# test job alert