+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- 109 Job alert subscription
 - 110 Job search alert notifications
 
 ## 4.0 (2020-02-24)